requirements.txt
chainlit
notebook
ipywidgets
tqdm
python-dotenv
openai
google-generativeai
langchain_google_genai

-e .